added authentication middleware
added auth middleware; adjusted postsController.js: checking permission on Post operations, like and remove like; adjusted post model: likes insted likesCount, addded authorId

+++ server/controllers/postsController.js
@@ -4,7 +4,7 @@ import Post from "../models/post.js";
 export const postsController = {
   async get(req, res) {
     try {
-      const data = await Post.find();
+      const data = await Post.find({}, { authorId: 0 });
       res.status(200).json(data);
     } catch (err) {
       res.status(err.status || 500).json({ message: err.message });
@@ -13,6 +13,8 @@ export const postsController = {
 
   async post(req, res) {
     const post = req.body;
+    const userId = req.headers.user.userId;
+    post.authorId = userId;
     const newPost = new Post(post);
     try {
       await newPost.save();
@@ -25,8 +27,10 @@ export const postsController = {
   async patch(req, res) {
     const post = req.body;
     const { id } = req.params;
+    const userId = req.headers.user.userId;
     try {
       await isIdValid(id);
+      await checkForPermissionOnPost(id, userId);
       const updatedPost = await Post.findByIdAndUpdate(id, post, {
         new: true,
       });
@@ -38,8 +42,10 @@ export const postsController = {
 
   async delete(req, res) {
     const { id } = req.params;
+    const userId = req.headers.user.userId;
     try {
       await isIdValid(id);
+      await checkForPermissionOnPost(id, userId);
       const deletedPost = await Post.findByIdAndRemove(id);
       res.status(200).json({
         message: `Post '${deletedPost.title}' successfully deleted.`,
@@ -51,14 +57,17 @@ export const postsController = {
 
   async likePost(req, res) {
     const { id } = req.params;
+    const userId = req.headers.user.userId;
     try {
       await isIdValid(id);
       const post = await Post.findById(id);
-      const updatedPost = await Post.findByIdAndUpdate(
-        id,
-        { likeCount: post.likeCount + 1 },
-        { new: true }
-      );
+      const index = post.likes.findIndex((id) => id === String(userId));
+      if (index === -1) {
+        post.likes.push(userId);
+      } else {
+        post.likes = post.likes.filter((id) => id !== String(userId));
+      }
+      const updatedPost = await Post.findByIdAndUpdate(id, post, { new: true });
       res.status(200).json(updatedPost);
     } catch (err) {
       res.status(err.status || 500).json({ message: err.message });
@@ -72,4 +81,11 @@ const isIdValid = async (id) => {
 
   if (!(await Post.findById(id)))
     throw { status: 400, message: "Update failed, post does not exist" };
+};
+
+const checkForPermissionOnPost = async (id, userId) => {
+  const oldPost = await Post.findById(id);
+  if (oldPost.authorId !== userId)
+    throw { status: 401, message: "No permission to perform any operation on this post." };
+  return;
 };
